Share currently broken configuration
This commit attempts to add C++ implementations of two
preprocessing functions (involved in scaleNotCenter, selectGenes).
However, the addition of 64bit word support in Armadillo for large
sparse matrices seems to break usage of a necessary BLAS routine
for optimizeALS on MacOS.

This configuration does work on R 3.5.3 on Debian 9 however.