Added an error catch to suggestK
When I was running through the tutorial I initially ran suggestK before running ScaleNotCenter (the command is mixed in with those from the 'precomputed list' set). It gave a fairly cryptic error. To save people in the future digging through the code to figure out what's wrong, probably best to just add this line. 

Thanks for developing this great package!